Refactor:타입 파일 => 폴더로 정리, 체험 리스트 성공 토스트 삭제
