+++ src/app/(usercard)/my-reservation/page.tsx
@@ -1,7 +1,7 @@
 export const dynamic = "force-dynamic";
 
 import API_URL from "@/constants/config";
-import { MyReservation } from "@/types/types";
+import { MyReservation } from "@/types/myReservationsTypes/myReservationsTypes";
 import MyReservationList from "@/app/(usercard)/my-reservation/MyReservationList/MyReservationList";
 import { customFetch } from "@/utils/customFetch";
